refactored data importing to skip journey insertion if bike station do not exists

+++ backend/src/database/initialisation.ts
@@ -22,9 +22,10 @@ export const initialiseBikeStationData = async () => {
 
   const queryBuilder = dataSource.createQueryBuilder();
 
+  let insertionCount = 0;
+
   const csvInsertions = bikeStationCsvFilepaths.map((csvPath) => {
     const parser = new CsvParser(csvPath, BIKE_STATION_COLUMN_HEADERS, bikeStationMapFn);
-    let insertionCount = 0;
     return parser.parse({
       bufferSize: AppConfig.DATA_IMPORT_BUFFER_SIZE,
       bufferProcessor: async (bikeStations) => {
@@ -37,6 +38,8 @@ export const initialiseBikeStationData = async () => {
 
   await Promise.all(csvInsertions);
   await dataSource.destroy();
+
+  return insertionCount;
 };
 
 export const initialiseJourneyData = async () => {
@@ -52,9 +55,10 @@ export const initialiseJourneyData = async () => {
 
   const queryBuilder = dataSource.createQueryBuilder();
 
+  let insertionCount = 0;
+
   const csvInsertions = journeyCsvFilepaths.map((csvPath) => {
     const parser = new CsvParser(csvPath, JOURNEY_COLUMN_HEADERS, journeyMapFn);
-    let insertionCount = 0;
     return parser.parse({
       bufferSize: AppConfig.DATA_IMPORT_BUFFER_SIZE,
       bufferProcessor: async (journeys) => {
@@ -71,21 +75,25 @@ export const initialiseJourneyData = async () => {
   //       Due to potentially huge datasets, individual insertion with row skips in case of foreign key violaton is too slow.
   //       So far I haven't figured out a way to batch insert such that journeys with invalid foreign keys would simply be skipped.
 
-  const existingBikeStationIds = (await queryBuilder.select('id').from(BikeStation, '').execute()).map(
-    (result: unknown) => (result && typeof result === 'object' && 'id' in result ? result.id : -1),
-  );
+  const existingBikeStationIds: Array<number> = (
+    await queryBuilder.from(BikeStation, 'station').select('station.id').getRawMany()
+  ).map((row) => row['station_id']);
 
-  const deleteResult = await queryBuilder
-    .delete()
-    .from(Journey, 'journey')
-    .where('return_station_id not in (:...ids) or departure_station_id not in (:...ids)', {
-      ids: existingBikeStationIds,
-    })
-    .execute();
+  if (existingBikeStationIds.length > 0) {
+    const deleteResult = await queryBuilder
+      .delete()
+      .from(Journey, 'journey')
+      .where('return_station_id not in (:...ids) or departure_station_id not in (:...ids)', {
+        ids: existingBikeStationIds,
+      })
+      .execute();
 
-  console.log(`Removed ${deleteResult.affected} journeys due to non-existing departure or return bike station`);
+    console.log(`Removed ${deleteResult.affected} journeys due to non-existing departure or return bike station`);
+  }
 
   await dataSource.destroy();
+
+  return insertionCount;
 };
 
 export const initialiseDatabase = async () => {
@@ -94,17 +102,22 @@ export const initialiseDatabase = async () => {
   const bikeStationsInitialised = (await AppDataSource.getRepository(BikeStation).find({ take: 1 })).length === 1;
   const journeysInitialised = (await AppDataSource.getRepository(Journey).find({ take: 1 })).length === 1;
 
+  let bikeStationInsertionCount = 0;
+  let journeyInsetionCount = 0;
+
   if (bikeStationsInitialised) {
     console.log('Bike stations already imported');
   } else {
-    await initialiseBikeStationData();
-    console.log('Bike stations initialised successfully');
+    bikeStationInsertionCount = await initialiseBikeStationData();
+    console.log(`${bikeStationInsertionCount} bike stations imported successfully`);
   }
 
   if (journeysInitialised) {
     console.log('Journeys already imported');
+  } else if (bikeStationsInitialised === false && bikeStationInsertionCount === 0) {
+    console.log('No bike stations exist to connect the journeys to, skipping journey insertion');
   } else {
-    await initialiseJourneyData();
-    console.log('Journeys initialised successfully');
+    journeyInsetionCount = await initialiseJourneyData();
+    console.log(`${journeyInsetionCount} journeys imported successfully`);
   }
 };
